[loader] Spin out image hashing out of image.c
Move MonoLoadedImages support code to loaded-images.c loaded-images-global.c
and loaded-images-netcore.c

Also spin out assembly-load-context.c for MonoAssemblyLoadContext functions
